backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
PyJWT
bcrypt
pydantic[email]
pydantic-settings
python-dotenv
python-multipart

# Testing
pytest
httpx
